Removed all warnings from exrImage.cpp. Made DWA attribute a float value.

+++ mrViewer/src/core/mrvImageOpts.h
@@ -35,11 +35,11 @@ class EXROpts : public ImageOpts
   protected:
     Imf::Compression _compression;
     Imf::PixelType   _pixel_type;
-    unsigned         _dwa_compression_level;
+    float            _dwa_compression_level;
   public:
     EXROpts() : _compression( Imf::ZIPS_COMPRESSION ),
                 _pixel_type( Imf::HALF ),
-                _dwa_compression_level( 45 )
+                _dwa_compression_level( 45.0f )
     {
     }
 
@@ -49,8 +49,8 @@ class EXROpts : public ImageOpts
     Imf::PixelType pixel_type() const { return _pixel_type; }
     void pixel_type( Imf::PixelType p ) { _pixel_type = p; }
 
-    unsigned compression_level() const { return _dwa_compression_level; }
-    void compression_level( unsigned p ) { _dwa_compression_level = p; }
+    float compression_level() const { return _dwa_compression_level; }
+    void compression_level( float p ) { _dwa_compression_level = p; }
 
 };
